Decoupled quikfynd core binaries from this rpm.
Earlier, this rpm assumed it would contain all quikfynd binaries
inside libraries folder. Now, a dependency has been created in the
rpm spec in this project, to install quikfynd.rpm from elsewhere.

I tested it with

rpm -Uvh *

after copying quikfynd-2.0.1-1.x86_64.rpm in the same folder as

app-quikfynd-2.0.1-1.quikfynddev.x86_64.rpm
app-quikfynd-core-2.0.1-1.quikfynddev.x86_64.rpm

With that setup, all the rpms installed on clearos VM and the
daemon started running. Note that the systemd specs are in this
project and not quikfynd.rpm. Since the init system is different
for different platforms, it would be better to keep that logic
in the clearos specific rpms. If one were to install only the
quikfynd.rpm separately, it would need to be started manually.

Signed-off-by: Karthik Veeramani <karthik@quikfynd.com>